+++ examples/server_profile_templates.py
@@ -103,6 +103,12 @@ if oneview_client.api_version >= 300:
         template["uri"], server_hardware_type_uri_for_transformation, enclosure_group_uri_for_transformation)
     pprint(transformation)
 
+#Get available networks
+print("\nGet available networks")
+available_networks = oneview_client.server_profile_templates.get_available_networks(
+    enclosureGroupUri=enclosure_group_uri, serverHardwareTypeUri=server_hardware_type_uri)
+print(available_networks)
+
 # Delete the created template
 print("\nDelete the created template")
 oneview_client.server_profile_templates.delete(basic_template)