+++ hpOneView/resources/servers/server_profile_templates.py
@@ -29,7 +29,6 @@ from future import standard_library
 
 standard_library.install_aliases()
 
-
 from hpOneView.resources.resource import ResourceClient
 
 
@@ -57,14 +56,15 @@ class ServerProfileTemplate(object):
     DEFAULT_VALUES = {
         '200': {'type': 'ServerProfileTemplateV1'},
         '300': {'type': 'ServerProfileTemplateV2'},
-        '500': {'type': 'ServerProfileTemplateV3'}
+        '500': {'type': 'ServerProfileTemplateV3'},
+        '600': {'type': 'ServerProfileTemplateV4'}
     }
 
     def __init__(self, con):
         self._connection = con
         self._client = ResourceClient(con, self.URI)
 
-    def get_all(self, start=0, count=-1, filter='', sort=''):
+    def get_all(self, start=0, count=-1, filter='', sort='', scope_uris=''):
         """
         Gets a list of server profile templates based on optional sorting and filtering and is constrained by start and
         count parameters.
@@ -89,12 +89,14 @@ class ServerProfileTemplate(object):
             sort:
                 The sort order of the returned data set. By default, the sort order is based
                 on create time with the oldest entry first.
+            scope_uris:
+                An expression to restrict the resources returned according to the scopes to which they are assigned.
 
         Returns:
             list: A list of server profile templates.
 
         """
-        return self._client.get_all(start=start, count=count, filter=filter, sort=sort)
+        return self._client.get_all(start=start, count=count, filter=filter, sort=sort, scope_uris=scope_uris)
 
     def get(self, id_or_uri):
         """
@@ -134,7 +136,7 @@ class ServerProfileTemplate(object):
         """
         return self._client.get_by_name(name)
 
-    def create(self, resource, timeout=-1):
+    def create(self, resource, timeout=-1, force=True):
         """
         Creates a server profile template.
 
@@ -143,14 +145,17 @@ class ServerProfileTemplate(object):
             timeout:
                 Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
                 in OneView, just stop waiting for its completion.
+            force: If set to true, the operation will ignore warnings for SAN storage.
 
         Returns:
             dict: Created resource.
 
         """
-        return self._client.create(resource=resource, timeout=timeout, default_values=self.DEFAULT_VALUES)
+        uri = self.URI + "?force={0}".format(force)
+        return self._client.create(resource=resource, uri=uri, timeout=timeout,
+                                   default_values=self.DEFAULT_VALUES)
 
-    def update(self, resource, id_or_uri):
+    def update(self, resource, id_or_uri, force=True):
         """
         Allows a server profile template object to have its configuration modified. These modifications can be as
         simple as a name or description change or more complex changes around the networking configuration.
@@ -162,9 +167,10 @@ class ServerProfileTemplate(object):
         Returns:
             dict: The server profile template resource.
         """
-        return self._client.update(resource=resource, uri=id_or_uri, default_values=self.DEFAULT_VALUES)
+        return self._client.update(resource=resource, uri=id_or_uri,
+                                   default_values=self.DEFAULT_VALUES, force=force)
 
-    def delete(self, resource, timeout=-1):
+    def delete(self, resource, timeout=-1, force=True):
         """
         Deletes a server profile template object from the appliance based on its profile template UUID.
 
@@ -177,7 +183,7 @@ class ServerProfileTemplate(object):
         Returns:
             bool: Indicates whether the resource was successfully deleted.
         """
-        return self._client.delete(resource=resource, timeout=timeout)
+        return self._client.delete(resource=resource, timeout=timeout, force=force)
 
     def get_new_profile(self, id_or_uri):
         """
@@ -217,4 +223,38 @@ class ServerProfileTemplate(object):
         """
         query_params = self.TRANSFORMATION_PATH.format(**locals())
         uri = self._client.build_uri(id_or_uri) + query_params
+        return self._client.get(id_or_uri=uri)
+
+    def get_available_networks(self, **kwargs):
+        """
+        Retrieves the list of Ethernet networks, Fibre Channel networks and network sets that are available
+        to a server profile template along with their respective ports. The scopeUris, serverHardwareTypeUri and
+        enclosureGroupUri parameters should be specified to get the available networks for a new server profile template.
+        The serverHardwareTypeUri, enclosureGroupUri, and profileTemplateUri should be specified to get available
+        networks for an existing server profile template.
+        The scopeUris parameter is ignored when the profileTemplateUri is specified.
+
+        Args:
+            enclosureGroupUri:
+                The URI of the enclosure group is required when the serverHardwareTypeUri specifies a blade server.
+            profileTemplateUri:
+                 If the URI of the server profile template is provided the list of available
+                 networks will include only networks that share a scope with the server profile template.
+            scopeUris:
+                An expression to restrict the resources returned according to the scopes to which they are assigned.
+            serverHardwareTypeUri:
+                 If the server hardware type specifies a rack server, the list of available network includes all
+                 networks that are applicable for the specified server hardware type. If the server hardware type
+                 specifies a blade server, the enclosureGroupUri parameter must be specified, and the list of
+                 available networks includes all networks that are applicable for the specified server hardware type
+                 and all empty bays within the enclosure group that can support the specified server hardware type.
+            view: The FunctionType (Ethernet or FibreChannel) to filter the list of networks returned.
+
+        Returns:
+            dict: The server profile resource.
+        """
+        query_string = '&'.join('{}={}'.format(key, value)
+                                for key, value in sorted(kwargs.items()) if value)
+        uri = self.URI + "{}?{}".format("/available-networks", query_string)
+
         return self._client.get(id_or_uri=uri)